chore: Build 86 baseline + carry-over uncommitted prep work
Snapshots the entire working tree state ahead of Build 87 (batch 2 feedback
fixes). Includes:

- The 5 Build 86 changes from the abstract-drifting-waffle plan:
  * Family member chooser sheet (regular vs expecting upstream of form)
  * Q6 water source icons for shared_well + not_sure
  * Q10 appliances reframe + Select all pill (new supportsSelectAll flag)
  * Q22 generator "same supplier?" confirmation card with Q19 lookup
  * Q25 garage restructure to 5 icon options + new Q25b EV charger split
- Plus a large amount of pre-existing uncommitted prep work that was sitting
  in the working tree before this session started: PropertyHook flow,
  PostQuizVendorDelegationSheet, QuizLocalContractorPicker, QuizContinueButton,
  ValuationRange, FindLocalVendorSheet, AccountCreationStep, multiple insurance
  / utility / town coverage migrations, find-local-vendors edge function,
  HouseQuiz feedback library, MaintenanceTaskReconciler updates, etc.

Committing as one snapshot per Tom's instruction so Build 87 has a clean
baseline to layer on top of without merge conflicts. Individual feature
commits will resume from the next change forward.

+++ Haven/Features/Property/Services/OnboardingScheduleGenerator.swift
@@ -7,7 +7,13 @@ struct PropertyLookupResult: Codable {
     var squareFootage: Int?
     var lotSize: Int?
     var bedrooms: Int?
-    var bathrooms: Int?
+    /// Phase 20 polish: ATTOM returns bathrooms as a Double (e.g. 3.5) to
+    /// represent half-baths. The previous `Int?` type caused the entire
+    /// JSON decode to fail with "Number 3.5 is not representable in Swift"
+    /// for any home with a half-bath, which silently dropped ALL property
+    /// data on a huge fraction of CT/NY HNW homes. Type widened to Double?
+    /// to match the upstream ATTOM contract.
+    var bathrooms: Double?
     var propertyType: String?
     let lastSaleDate: String?
     let lastSalePrice: Double?
